% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhs_individual_id}
\alias{dhs_individual_id}
\title{Create individual ID for DHS surveys}
\usage{
dhs_individual_id(cluster, household, line)
}
\arguments{
\item{cluster}{Cluster ID}

\item{household}{Household ID}

\item{line}{Line number}
}
\description{
Create individual ID for DHS surveys
}
