% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhscc_to_iso3}
\alias{dhscc_to_iso3}
\title{Convert DHS country code to ISO3C code}
\usage{
dhscc_to_iso3(dhscc)
}
\arguments{
\item{dhscc}{Two letter DHS country code}
}
\description{
Convert DHS country code to ISO3C code
}
