% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{extract_survey_vars}
\alias{extract_survey_vars}
\title{Extract survey variables}
\usage{
extract_survey_vars(df, survey_id_c, variable_recode, dataset_type, analysis_c)
}
\arguments{
\item{df}{Raw survey dataframe}

\item{survey_id_c}{Survey ID}

\item{variable_recode}{Variable recode book}

\item{dataset_type}{Dataset type e.g.....}

\item{analysis}{One of "circ", "sexbehav" etc}
}
\description{
Extract survey variables
}
