% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{recode_mics_id_vars}
\alias{recode_mics_id_vars}
\title{Recode ID variables in MICS surveys}
\usage{
recode_mics_id_vars(variable_recode, survey_id_c, dataset_type)
}
\arguments{
\item{variable_recode}{Variable recode book}

\item{survey_id_c}{Survey ID}

\item{dataset_type}{Dataset type e.g.....}
}
\description{
MICS surveys do not have consistent cluster, household, and
line number variables. This standardises them.
}
