% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{recode_survey_variables}
\alias{recode_survey_variables}
\title{Recode survey variables}
\usage{
recode_survey_variables(
  df,
  survey_id_c,
  value_recode,
  dataset_type,
  analysis_c
)
}
\arguments{
\item{df}{Survey dataset}

\item{survey_id_c}{Survey ID}

\item{value_recode}{Value recode excel sheet}

\item{dataset_type}{Dataset type e.g.....}

\item{analysis}{One of "circ", "sexbehav" etc}
}
\description{
Recode survey variables
}
