% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_funs.R
\name{val_recode}
\alias{val_recode}
\title{Recode survey column}
\usage{
val_recode(col, col_name, survey_id_c, dataset_type, analysis_c)
}
\arguments{
\item{col}{Column passed from across()}

\item{col_name}{Column name passed by dplyr::cur_column()}

\item{survey_id_c}{Survey ID}

\item{dataset_type}{Dataset type e.g.....}

\item{analysis}{One of "circ", "sexbehav" etc}
}
\description{
Recode survey column
}
